+++ package/distance/points.go
@@ -1,8 +1,8 @@
 package distance
 
 type Point struct {
-	X float64
-	Y float64
+	X float32
+	Y float32
 }
 
 type CoordinationType uint8
